chore(chainConfig): consolidate use of RPC url, merge duplicated wagmi config files, single source of truth for chain config
